Explicitly cast dev and inode to u64 in Unix-specific functions

+++ src/pipeline.rs
@@ -914,7 +914,7 @@ fn is_dot_or_dotdot(name: &[u8]) -> bool {
 
 #[cfg(unix)]
 fn dev_inode_from_stat(st: &libc::stat) -> (u64, u64) {
-    (st.st_dev, st.st_ino)
+    (st.st_dev as u64, st.st_ino as u64)
 }
 
 #[cfg(unix)]
